refactor all packet related or socket related code to separate classes
